DE1718 Fix Wrong Image In Product Feed
The collection was missing the 'image' attribute in the
fields to select causing the product to not have image loaded
and cause default place holder image to be used when generating
the feed. Writing unittest as integration with fixtures to
prove that a product loaded with a image attribute
will extract out the valid url path to the image.
However, a product that has not image attribute or
image attribute of null will extract the place holder
for the product.